Switch position statement to use Uint32
Position is restricted to 32bit unsigned, there is just no point
to say it is a Long.

JIRA: YANGTOOLS-1065

+++ yang/yang-parser-rfc7950/src/main/java/org/opendaylight/yangtools/yang/parser/rfc7950/stmt/position/EmptyPositionStatement.java
@@ -7,8 +7,10 @@
  */
 package org.opendaylight.yangtools.yang.parser.rfc7950.stmt.position;
 
+import org.opendaylight.yangtools.yang.common.Uint32;
+
 final class EmptyPositionStatement extends AbstractPositionStatement {
-    EmptyPositionStatement(final Long argument) {
+    EmptyPositionStatement(final Uint32 argument) {
         super(argument);
     }
 }
